+++ packages/plugin-text/src/controls/index.tsx
@@ -1,92 +1,63 @@
 import * as React from 'react'
 import { Editor } from 'slate'
 import { EditorProps } from 'slate-react'
-import {
-  toggleEmphasize,
-  isEmphasized,
-  toggleStrong,
-  isStrong
-} from '../plugins/rich-text'
-import { Button } from '../toolbar/button'
 import { TextPlugin } from '..'
-import { isLink, unwrapLink, wrapLink } from '../plugins/link'
-import {
-  Icon,
-  faLink,
-  faBold,
-  faItalic,
-  HoveringOverlay,
-  BottomToolbar
-} from '@edtr-io/editor-ui'
-import { OverlayContext } from '@edtr-io/core'
-import { insertKatex, isKatex } from '../plugins/katex'
+import { HoveringOverlay, BottomToolbar } from '@edtr-io/editor-ui'
+import { SlatePluginClosure } from '../factory/types'
+import { DefaultControls } from './default'
+import { HeadingControls } from './headings'
 
-export const Controls: React.FunctionComponent<{
+export enum VisibleControls {
+  All,
+  Headings
+}
+
+export interface ControlProps {
   editor: Editor
+  name: string
   readOnly?: boolean
+}
+
+export interface UiPluginOptions {
+  Component: React.ComponentType<Partial<EditorProps> & ControlProps>
+}
+
+const ControlsSwitch: React.FunctionComponent<{
+  editor: Editor
+  name: string
 }> = props => {
-  const { editor } = props
-  const overlayContext = React.useContext(OverlayContext)
+  const [visibleControls, setVisibleControls] = React.useState(
+    VisibleControls.All
+  )
+  switch (visibleControls) {
+    case VisibleControls.All:
+      return <DefaultControls {...props} switchControls={setVisibleControls} />
+    case VisibleControls.Headings:
+      return <HeadingControls {...props} switchControls={setVisibleControls} />
+  }
+}
+
+export const Controls: React.FunctionComponent<ControlProps> = props => {
+  const selectionCollapsed = props.editor.value.selection.isCollapsed
   return (
     <React.Fragment>
-      {editor.value.selection.isCollapsed &&
-      ((!isStrong(editor) && !isEmphasized(editor)) ||
-        isLink(editor)) ? null : (
+      {!selectionCollapsed && (
         <HoveringOverlay position={'above'}>
-          <Button
-            active={isStrong(editor)}
-            onClick={() => toggleStrong(editor).focus()}
-          >
-            <Icon icon={faBold} />
-          </Button>
-          <Button
-            active={isEmphasized(editor)}
-            onClick={() => toggleEmphasize(editor).focus()}
-          >
-            <Icon icon={faItalic} />
-          </Button>
-          <Button
-            active={isLink(editor)}
-            onClick={() =>
-              isLink(editor)
-                ? unwrapLink(editor).focus()
-                : wrapLink()(editor, overlayContext)
-            }
-          >
-            <Icon icon={faLink} />
-          </Button>
+          <ControlsSwitch {...props} />
         </HoveringOverlay>
       )}
-      {!props.readOnly && (
+      {!props.readOnly && selectionCollapsed && (
         <BottomToolbar>
-          <Button
-            active={false}
-            onClick={() => {
-              const active = isKatex(editor)
-
-              if (!active) {
-                insertKatex(editor)
-              }
-            }}
-          >
-            f(x)
-          </Button>
+          <ControlsSwitch {...props} />
         </BottomToolbar>
       )}
     </React.Fragment>
   )
 }
 
-export interface UiPluginOptions {
-  Component: React.ComponentType<
-    Partial<EditorProps> & {
-      editor: Editor
-      readOnly?: boolean
-    }
-  >
-}
-
-export const createUiPlugin = (options: UiPluginOptions) => (): TextPlugin => {
+export const createUiPlugin = (options: UiPluginOptions) => (
+  pluginClosure: SlatePluginClosure
+): TextPlugin => {
   const { Component } = options
 
   return {
@@ -95,10 +66,13 @@ export const createUiPlugin = (options: UiPluginOptions) => (): TextPlugin => {
       if (readOnly) {
         editor.blur()
       }
+      const name = pluginClosure.current ? pluginClosure.current.name : ''
       const children = next()
       return (
         <React.Fragment>
-          {!readOnly ? <Component editor={editor} {...props} /> : null}
+          {!readOnly ? (
+            <Component editor={editor} {...props} name={name} />
+          ) : null}
           {children}
         </React.Fragment>
       )