+++ packages/plugin-text/src/plugins/index.ts
@@ -3,10 +3,12 @@ import { createParagraphPlugin } from './paragraph'
 import { createLinkPlugin } from './link'
 import { TextPluginOptions } from '../factory/types'
 import { createKatexPlugin } from './katex'
+import { createHeadingsPlugin } from './headings'
 
 export const plugins: TextPluginOptions['plugins'] = [
   createParagraphPlugin(),
   createRichTextPlugin(),
   createLinkPlugin(),
-  createKatexPlugin()
+  createKatexPlugin(),
+  createHeadingsPlugin()
 ]