+++ scripts/changelog.ts
@@ -180,7 +180,12 @@ async function exec(): Promise<void> {
       ],
       added: [
         '**plugin-h5p**. Add H5p.com plugin',
-        '**plugin-image**. Add MaxWidth for images'
+        '**plugin-image**. Add MaxWidth for images',
+        '**plugin-rows**. Added more theming options',
+        '**core**. Added `isEmpty(state: State, id: string)` to check if a plugin state is empty',
+        '**core**. Added optional function `isEmpty: (state: StateDescriptorValueType<S>) => boolean` to plugin definition',
+        '**plugin-text**. Added visual latex editor',
+        '**plugin-text**. Added headings plugin'
       ],
       changed: [
         '**plugin-text**. Hide placeholder in render mode',
@@ -192,7 +197,8 @@ async function exec(): Promise<void> {
         '**plugin-rows**. Added more theming options',
         '**core**. Added `isEmpty(state: State, id: string)` to check if a plugin state is empty',
         '**core**. Added optional function `isEmpty: (state: StateDescriptorValueType<S>) => boolean` to plugin definition',
-        '**plugin-text**. Added visual latex editor'
+        '**plugin-text**. Added visual latex editor',
+        '**plugin-rows**. Improved styling of add menu and controls'
       ],
       fixed: ['**plugin-image**. Hide config-overlay in render mode'],
       internal: [